requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-core
langchain-text-splitters
langchain-community
langchain-google-genai==1.0.10
boto3
tiktoken
faiss-cpu